HOOK's-refactor
